LPS-77925 Move new upgrade from core to journal-service module
